requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pytelegrambotapi==4.14.0
requests==2.31.0
atlassian-python-api==3.41.9
elasticsearch==8.11.0
structlog==23.2.0
python-multipart==0.0.6  # for FastAPI file uploads if needed
prometheus-client==0.19.0